+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Major refactoring of the source code.
 ### Upgraded
 - `keyple-demo-common-lib:1.0.0-SNAPSHOT`
-- `calypsonet-terminal-calypso-java-api:1.2.+`
+- `calypsonet-terminal-calypso-java-api:1.4.+`
 - `keyple-service-java-lib:2.1.0`
 - `keyple-service-resource-java-lib:2.0.2`
 - `keyple-card-calypso-java-lib:2.2.1`